+++ src/core/univ3forks/Bulla.sol
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.25;
+
+address constant bullaFactory = 0x425EC3de5FEB62897dbe239Aa218B2DC035DCDF1;
+uint8 constant bullaForkId = 31;