+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Master list of UniV3 forks:
   27. SpookySwap
   28. Wagmi
   29. SwapX (Algebra-like)
+  30. KodiakV3
+  31. Bulla Exchange (Algebra-like)
 
 ---
 
@@ -49,6 +51,8 @@ Master list of UniV3 forks:
   * Add UniswapV3 UniV3 fork to Unichain
   * Add UniswapV4 actions to Unichain
 * Deploy Settler to Berachain network
+  * Add KodiakV3 UniV3 fork to Berachain
+  * Add Bulla Exchange UniV3 (Algebra style) fork to Berachain
 
 ## 2025-01-23